+++ lib/features/player/player_actions.dart
@@ -9,6 +9,7 @@ import 'package:shared_preferences/shared_preferences.dart';
 
 import '../../core/api/server_adapter.dart';
 import '../../core/cache/cache_manager.dart';
+import '../../core/platform/app_platform.dart';
 import '../../core/platform/local_fs.dart';
 import '../../core/download/download_service.dart';
 import '../../core/local/local_library.dart';